Complete overhaul
I re-did the entire app. Here's a summary.

Code changes:
- Rebuilt using PHP framework (Kohana).
- The old version's cron job for scraping the feeds saved the results as
a table in a database. Now, instead it copies the XML or JSON files to
the server for the app to parse later. There is no database involved
anymore, and this way the live feed is parsed the same way as the saved
feeds.

Changes for users:
- The marker colors show how full the stations are.
- Markers with colored borders (strokes) indicate that the station is
filling up or emptying quickly. This is based on a formula using data
the app has scraped from the past half hour.
- Users can update all the station data without refreshing the page. Two
ways to do so: a button that's always visible in the top right corner,
or a link in each marker's popup box.