Handle requests for non-existent station info
On individual-station pages, if the station isn't in one of the recorded feeds,
don't try to find its info in that feed. (Previously in this situation, the code
would inadvertently use the first station in the array if the desired station
didn't exist.) If the station isn't in the live feed, redirect to the station list page.